pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["baktiyar <baktiarzaksylykov22@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.116.1"
pydantic-settings = "^2.10.1"
asyncpg = "^0.30.0"
alembic = "^1.16.4"
pydantic = {extras = ["email"], version = "^2.11.7"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.35.0"
pyjwt = "^2.10.1"
bcrypt = "4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
